wip: break down table components
