+++ packages/frontend/src/components/Explorer/ResultsCard/ExplorerResults2.tsx
@@ -15,7 +15,8 @@ import React, { useMemo } from 'react';
 import styled from 'styled-components';
 import { useExplore } from '../../../hooks/useExplore';
 import { useExplorer } from '../../../providers/ExplorerProvider';
-import Table, { HeaderProps, TableColumn } from '../../common/Table';
+import Table from '../../common/Table';
+import { HeaderProps, TableColumn } from '../../common/Table/types';
 import { CellContextMenu } from '../../ResultsTable/CellContextMenu';
 import ColumnHeaderContextMenu from '../../ResultsTable/ColumnHeaderContextMenu';
 import TableCalculationHeaderButton from '../../TableCalculationHeaderButton';
