+++ packages/frontend/src/components/ResultsTable/ColumnHeaderContextMenu.tsx
@@ -11,7 +11,7 @@ import { useFilters } from '../../hooks/useFilters';
 import { useExplorer } from '../../providers/ExplorerProvider';
 import { useTracking } from '../../providers/TrackingProvider';
 import { EventName } from '../../types/Events';
-import { HeaderProps, TableColumn } from '../common/Table';
+import { HeaderProps, TableColumn } from '../common/Table/types';
 
 const ColumnHeaderContextMenu: React.FC<HeaderProps> = ({
     children,
