+++ packages/frontend/src/components/common/Table/ScrollableTable/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Table, TableScrollableWrapper } from '../Table.styles';
+import TableBody from './TableBody';
+import TableFooter from './TableFooter';
+import TableHeader from './TableHeader';
+
+const ScrollableTable = () => {
+    return (
+        <TableScrollableWrapper>
+            <Table bordered condensed>
+                <TableHeader />
+                <TableBody />
+                <TableFooter />
+            </Table>
+        </TableScrollableWrapper>
+    );
+};
+
+export default ScrollableTable;
