+++ packages/frontend/src/components/common/Table/TableProvider.tsx
@@ -1,0 +1,88 @@
+import {
+    ColumnOrderState,
+    getCoreRowModel,
+    getPaginationRowModel,
+    Table,
+    useReactTable,
+} from '@tanstack/react-table';
+import React, { createContext, FC, useContext } from 'react';
+import {
+    CellContextMenuProps,
+    HeaderProps,
+    ROW_NUMBER_COLUMN_ID,
+    TableColumn,
+    TableRow,
+} from './types';
+
+type Props = {
+    data: TableRow[];
+    columns: TableColumn[];
+    headerContextMenu?: FC<HeaderProps>;
+    headerButton?: FC<HeaderProps>;
+    cellContextMenu?: FC<CellContextMenuProps>;
+};
+
+type TableContext = {
+    data: TableRow[];
+    table: Table<TableRow>;
+    headerContextMenu?: FC<HeaderProps>;
+    headerButton?: FC<HeaderProps>;
+    cellContextMenu?: FC<CellContextMenuProps>;
+};
+
+const Context = createContext<TableContext | undefined>(undefined);
+
+const rowColumn: TableColumn = {
+    id: ROW_NUMBER_COLUMN_ID,
+    header: '#',
+    cell: (props) => props.row.index + 1,
+    footer: 'Total',
+    meta: {
+        width: 30,
+    },
+};
+
+export const TableProvider: FC<Props> = ({
+    data,
+    columns,
+    headerButton,
+    headerContextMenu,
+    cellContextMenu,
+    children,
+}) => {
+    const [columnVisibility, setColumnVisibility] = React.useState({});
+    const [columnOrder, setColumnOrder] = React.useState<ColumnOrderState>([]);
+    const table = useReactTable({
+        data,
+        columns: [rowColumn, ...columns],
+        state: {
+            columnVisibility,
+            columnOrder: [...new Set([ROW_NUMBER_COLUMN_ID, ...columnOrder])],
+        },
+        onColumnVisibilityChange: setColumnVisibility,
+        onColumnOrderChange: setColumnOrder,
+        getCoreRowModel: getCoreRowModel(),
+        getPaginationRowModel: getPaginationRowModel(),
+    });
+    return (
+        <Context.Provider
+            value={{
+                data,
+                table,
+                headerButton,
+                headerContextMenu,
+                cellContextMenu,
+            }}
+        >
+            {children}
+        </Context.Provider>
+    );
+};
+
+export function useTableContext(): TableContext {
+    const context = useContext(Context);
+    if (context === undefined) {
+        throw new Error('useTableContext must be used within a TableProvider');
+    }
+    return context;
+}
